requirements.txt
pybit
requests
python-dotenv
